% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestClusteriHMM.R
\name{bestClusteriHMM}
\alias{bestClusteriHMM}
\title{Identify best clustering of states in PSBP-iHMM model}
\usage{
bestClusteriHMM(ihmm)
}
\arguments{
\item{ihmm}{object of type "ihmm"}
}
\value{
best clustering of z
}
\description{
Identify best clustering of states in PSBP-iHMM model
}
