% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZminus1.R
\name{detZminus1}
\alias{detZminus1}
\title{Determine possible states and previous states}
\usage{
detZminus1(i, sli, pizi, ui, t.max)
}
\arguments{
\item{i}{sampling day}

\item{sli}{state list}

\item{pizi}{pi list}

\item{ui}{u list}

\item{t.max}{length of time series}
}
\value{
list of current and previous possible states
}
\description{
Determine possible states and previous states
}
