% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZminus1forNone.R
\name{detZminus1forNone}
\alias{detZminus1forNone}
\title{Determine possible states when no covariates}
\usage{
detZminus1forNone(i, sli, piz, ui, t.max)
}
\arguments{
\item{i}{sampling day}

\item{sli}{state list}

\item{ui}{u list}

\item{t.max}{length of time series}

\item{pizi}{pi list}
}
\value{
list of current and previous possible states
}
\description{
Determine possible states when no covariates
}
