% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDPMM.R
\name{fitDPMM}
\alias{fitDPMM}
\title{Fit DPMM with stick breaking process}
\usage{
fitDPMM(
  niter,
  nburn,
  y,
  ycomplete = NULL,
  priors = NULL,
  K.start = NULL,
  z.true = NULL,
  lod = NULL,
  mu.true = NULL,
  missing = FALSE,
  tau2 = NULL,
  a.tune = NULL,
  b.tune = NULL,
  resK = FALSE,
  eta.star = NULL,
  len.imp = NULL
)
}
\arguments{
\item{niter}{number of iterations}

\item{nburn}{burn-in}

\item{y}{list of time data for each time series}

\item{ycomplete}{complete data for evaluating imputation}

\item{priors}{list of priors}

\item{K.start}{starting value number of states}

\item{z.true}{list of true hidden states for error measure}

\item{lod}{vector of lower limits of detection for p exposures}

\item{mu.true}{matrix of true exposure means for each true state}

\item{missing}{logical; if TRUE then the data set contains missing data, default is FALSE}

\item{tau2}{tuning parameter for MH algorithm for reparameterized Sigma}

\item{a.tune}{tuning parameter for MH algorithm for reparameterized Sigma}

\item{b.tune}{tuning parameter for MH algorithm for reparameterized Sigma}

\item{resK}{logical; resolvent kernel for MH update a}

\item{eta.star}{reslvent kernel parameter for MH update a}

\item{len.imp}{number of imputations to save}
}
\value{
list of results
}
\description{
Fit DPMM with stick breaking process
}
