% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedK_clustering.R
\name{fixedK_clustering}
\alias{fixedK_clustering}
\title{MVN model based clustering}
\usage{
fixedK_clustering(
  niter,
  nburn,
  y,
  ycomplete = NULL,
  priors = NULL,
  K.true,
  z.true,
  lod = NULL,
  missing = FALSE,
  gibbs_update = TRUE,
  tau2 = NULL,
  a.tune = NULL,
  b.tune = NULL,
  resK = FALSE,
  eta.star = NULL,
  len.imp = NULL,
  holdout = NULL
)
}
\arguments{
\item{niter}{number of total iterations}

\item{nburn}{number of burn-in iterations}

\item{y}{list of time series data for each time series}

\item{ycomplete}{complete data, if available, for evaluating imputations}

\item{priors}{list of priors}

\item{K.true}{specified number of hidden states}

\item{z.true}{specified hidden state trajectories}

\item{lod}{limit of detection}

\item{missing}{indicator if there is missing data}

\item{gibbs_update}{logical; if TRUE implement Gibbs sampling of state-specific covariance matrices, otherwise, implement MH updates of reparameterized Sigma}

\item{tau2}{variance tuning parameter for normal proposal in MH update of lower triangular elements in decomposition of Sigma}

\item{a.tune}{shape tuning parameter for inverse gamma proposal in MH update of diagonal elements in decomposition of Sigma}

\item{b.tune}{rate tuning parameter for inverse gamma proposal in MH update of diagonal elements in decomposition of Sigma}

\item{resK}{logical; if TRUE a resolvent kernel is used in MH update for lower triangular elements in decomposition of Sigma}

\item{eta.star}{resolvent kernel parameter, must be a real value greater than 1. In the resolvent kernel we take a random draw from the geometric distribution with mean (1-p)/p, eta.star = 1/p.}

\item{len.imp}{number of imputations to save. Imputations will be taken at equally spaced iterations between nburn and niter.}

\item{holdout}{list of indicators of missing type in holdout data set, 0 = observed, 1 = MAR, 2 = below LOD, for imputation validation purposes}
}
\value{
list
}
\description{
Function to fit MVN models to data sub-setted by known states
}
