% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndepSimResults.R
\name{getIndepSimRsults}
\alias{getIndepSimRsults}
\title{Summarize simulation results from indpendently fit iHMMs}
\usage{
getIndepSimRsults(fit1, y)
}
\arguments{
\item{fit1}{list of objects of type iHMM}

\item{y}{observed data}
}
\value{
a list with components 
\itemize{
    \item hamming: hamming distance
    \item mu.mse: MSE for state-specific means
    \item mar.mse: MSE for MAR imputations 
    \item lod.mse: MSE for imputations below LOD
}
}
\description{
Summarize simulation results from indpendently fit iHMMs
}
