% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJointSimResults.R
\name{getJointSimResults}
\alias{getJointSimResults}
\title{Summarize simulation results from jointly fit iHMMs}
\usage{
getJointSimResults(fit1)
}
\arguments{
\item{fit1}{object of type iHMM}
}
\value{
a list with components 
\itemize{
    \item hamming: hamming distance
    \item mu.mse: MSE for state-specific means
    \item mar.mse: MSE for MAR imputations 
    \item lod.mse: MSE for imputations below LOD
}
}
\description{
Summarize simulation results from jointly fit iHMMs
}
