% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamdist.R
\name{hamdist}
\alias{hamdist}
\title{Greedily assign estimated states to true states then calculate Hamming distance}
\usage{
hamdist(z1, z2)
}
\arguments{
\item{z1}{vector of true states}

\item{z2}{vector of estimated states}
}
\value{
hamming distance for most overlap
}
\description{
Greedily assign estimated states to true states then calculate Hamming distance
}
