% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismissing.R
\name{ismissing}
\alias{ismissing}
\title{function to indicate missing values}
\usage{
ismissing(x)
}
\arguments{
\item{x}{data}
}
\value{
matrix same size as x, recoded to denote missing type. 0 = observed, 1 = NA (missing at random), 2 = -Inf (below LOD)
}
\description{
function to indicate missing values
}
