% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mciHMM.R
\name{mciHMM}
\alias{mciHMM}
\title{Fit mc-iHMM with covariates and/or repeated measures}
\usage{
mciHMM(
  niter,
  nburn,
  y,
  rmlist,
  ycomplete = NULL,
  X,
  priors = NULL,
  K.start = NULL,
  z.true = NULL,
  lod = NULL,
  mu.true = NULL,
  missing = FALSE,
  tau2 = 0.1,
  a.tune = 10,
  b.tune = 1,
  resK = FALSE,
  eta.star = NULL,
  len.imp = 1
)
}
\arguments{
\item{niter}{number of iterations}

\item{nburn}{burn-in}

\item{y}{list of time data for each time series}

\item{rmlist}{vector identifying repeated measures}

\item{ycomplete}{complete data for evaluating imputation}

\item{X}{list of design matrix for each i}

\item{priors}{list of priors}

\item{K.start}{starting value number of states}

\item{z.true}{list of true hidden states for error measure}

\item{lod}{vector of lower limits of detection for p exposures}

\item{mu.true}{matrix of true exposure means for each true state}

\item{missing}{logical; if TRUE then the data set y contains missing data, default is FALSE}

\item{tau2}{tuning parameter for MH update for a}

\item{a.tune}{tuning parameter for MH udpate for lams}

\item{b.tune}{tuning parameter for MH update for lams}

\item{resK}{logical; resolvent kernel for MH update a}

\item{eta.star}{reslvent kernel parameter for MH update a}

\item{len.imp}{number of imputations to save}
}
\value{
list of results
}
\description{
Fit mc-iHMM with covariates and/or repeated measures
}
