% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgamma.R
\name{mgamma}
\alias{mgamma}
\title{multivariate gamma function}
\usage{
mgamma(nu, p)
}
\arguments{
\item{nu}{nu (Wishart df)}

\item{p}{p (dimension of Sigma; number of exposures)}
}
\value{
Gamma_p((nu+1)/2)/Gamma_p(nu/2)
}
\description{
multivariate gamma function
}
