% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMode.R
\name{minMode}
\alias{minMode}
\title{Min Mode: Find the least common value}
\usage{
minMode(x)
}
\arguments{
\item{x}{vector of states}
}
\value{
least common value
}
\description{
Min Mode: Find the least common value
}
