\name{psbpHMM-package}
\alias{psbpHMM-package}
\alias{psbpHMM}
\docType{package}
\title{
\packageTitle{psbpHMM}
}
\description{
\packageDescription{psbpHMM}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{psbpHMM}
\packageIndices{psbpHMM}
Overview of package. 
}
\author{
\packageAuthor{psbpHMM}

Maintainer: \packageMaintainer{psbpHMM}
}
\references{
literature 
}
\keyword{ package }

