% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate data for psbpHMM package}
\usage{
simdat(
  n,
  t.max = 24,
  K = 6,
  p = 3,
  tempTrend = TRUE,
  lodRemove = FALSE,
  marRemove = FALSE,
  lodmis = 0,
  marmis = 0,
  sf = 0.1
)
}
\arguments{
\item{n}{number of time series}

\item{t.max}{length of each time series}

\item{K}{number of clusters to be shared across all time series}

\item{p}{number of exposures}

\item{tempTrend}{logical; if true, all time series have a common temporal trend}

\item{lodRemove}{logical; if true, remove data emulating observations below the limit of detection}

\item{marRemove}{logical; if true, remove data emulating observations missing at random}

\item{lodmis}{percent of data to remove as below limit of detection}

\item{marmis}{percent of data to remove as missing at random}

\item{sf}{parameter controlling level of variation in state-specific covariance matrices}
}
\value{
a list with components
\itemize{
    \item y: scaled simulated data with missing observations set to NA for MAR or -Inf for below LOD 
    \item y.complete: scaled complete data 
    \item z.true: true hidden state trajectories for each time series
    \item K: true number of hiddens states
    \item mu.true: scaled true state-specific means
    \item Sigma.true: unscaled true state-specific covariance matrices
    \item lod: limit of detection
}
}
\description{
Simulate multiple time series data from a fixed-state MVN hidden markov model ∂
}
