% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate data to fit methods in the R package `psbpHMM'}
\usage{
simData(n = 20, t.max = 288, K = 6, p = 3, trend = "shared", missingLevel = 0)
}
\arguments{
\item{n}{number of time series}

\item{t.max}{length of each time series}

\item{K}{number of clusters to be shared across all time series}

\item{p}{number of exposures}

\item{trend}{"shared" (default) for shared temporal trends or "distinct" for distinct temporal trends among n time series}

\item{missingLevel}{missing data level, data will be removed and split between missing at random and below LOD}
}
\value{
a list with components
\itemize{
    \item y: scaled simulated data with missing observations set to NA for MAR or -Inf for below LOD 
    \item y.complete: scaled complete data 
    \item z.true: true hidden state trajectories for each time series
    \item K: true number of hiddens states
    \item mu.true: scaled true state-specific means
    \item lod: limit of detection
}
}
\description{
Simulate multiple time series data from a fixed-state MVN hidden Markov model
}
