% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAlphaJK.R
\name{updateAlphaJK}
\alias{updateAlphaJK}
\title{Update alpha_jk}
\usage{
updateAlphaJK(
  j,
  k,
  n,
  t.max,
  z,
  vinv.alpha,
  sig2inv.alpha,
  w.z,
  X,
  beta.k,
  beta.sk = NULL,
  m.alpha,
  mu.alpha
)
}
\arguments{
\item{j}{previous state}

\item{k}{current state}

\item{n}{number of time series}

\item{t.max}{length of time series}

\item{z}{hidden states}

\item{vinv.alpha}{prior}

\item{sig2inv.alpha}{prior}

\item{w.z}{w}

\item{X}{covariates}

\item{beta.k}{overall beta}

\item{beta.sk}{subject specific beta}

\item{m.alpha}{prior}

\item{mu.alpha}{prior}
}
\value{
sampled value for alpha_jk
}
\description{
Update alpha_jk
}
