% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateW.R
\name{updateW}
\alias{updateW}
\title{Update W with or without repeated measures}
\usage{
updateW(t, l, i, alpha.0k, X, beta.k, beta.sk = NULL, alpha.jk, z)
}
\arguments{
\item{t}{time point}

\item{l}{state}

\item{i}{time series}

\item{alpha.0k}{alpha0k}

\item{X}{covariates}

\item{beta.k}{overall beta}

\item{beta.sk}{subject specific beta}

\item{alpha.jk}{alphajk}

\item{z}{hidden states}
}
\value{
sampled value for W
}
\description{
Update W with or without repeated measures
}
