% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateZ_Rcode.R
\name{updateZ_Rcode}
\alias{updateZ_Rcode}
\title{Update hidden states}
\usage{
updateZ_Rcode(
  i,
  yi,
  mu,
  cholSigma,
  detR.stari,
  nu.df,
  K,
  log.stuff,
  t.max,
  detz,
  sli
)
}
\arguments{
\item{i}{sampling day}

\item{yi}{y}

\item{mu}{mu}

\item{cholSigma}{chol Sigma}

\item{detR.stari}{detRstar}

\item{nu.df}{nu df}

\item{K}{number of states}

\item{log.stuff}{log stuff constant}

\item{t.max}{length of time series}

\item{detz}{possible states for t-1}

\item{sli}{state lists}
}
\value{
updated states
}
\description{
Update hidden states
}
