% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upStateList_lapply.R
\name{upStateList_lapply}
\alias{upStateList_lapply}
\title{update state list}
\usage{
upStateList_lapply(i, u, pi.z, K, t.max)
}
\arguments{
\item{i}{sampling day}

\item{u}{list of u}

\item{pi.z}{list of pi}

\item{K}{number of states}

\item{t.max}{time series length}
}
\value{
updated state list
}
\description{
update state list
}
