% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upStateList_lapply_nox.R
\name{upStateList_lapply_nox}
\alias{upStateList_lapply_nox}
\title{update state list when no covariates}
\usage{
upStateList_lapply_nox(i, u, pi.z, K, t.max)
}
\arguments{
\item{i}{sampling day}

\item{u}{list of u}

\item{pi.z}{list of pi}

\item{K}{number of states}

\item{t.max}{time series length}
}
\value{
udpated state list
}
\description{
update state list when no covariates
}
