% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMinusb.R
\name{wMinusb}
\alias{wMinusb}
\title{Calculate W minus Xbeta}
\usage{
wMinusb(i, t, k, w.z, beta.k, beta.sk = NULL, X)
}
\arguments{
\item{i}{time series}

\item{t}{time point}

\item{k}{state}

\item{w.z}{w}

\item{beta.k}{overall beta}

\item{beta.sk}{subject specific beta}

\item{X}{covariates}
}
\value{
w minus X%*%beta
}
\description{
Calculate W minus Xbeta
}
