% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlso_wrapper.R
\name{dlso_wrapper}
\alias{dlso_wrapper}
\title{Identify best clustering of states in PSBP-iHMM model}
\usage{
dlso_wrapper(ihmm)
}
\arguments{
\item{ihmm}{object of type "ihmm"}
}
\value{
best clustering of z
}
\description{
Identify best clustering of states in PSBP-iHMM model
}
