% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedK_clustering.R
\name{fixedK_clustering}
\alias{fixedK_clustering}
\title{MVN model based clustering with fixed K and fixed state assignments}
\usage{
fixedK_clustering(
  niter,
  nburn,
  y,
  ycomplete = NULL,
  priors = NULL,
  K.true,
  z.true,
  lod = NULL,
  missing = FALSE,
  gibbs_update = TRUE,
  tau2 = NULL,
  a.tune = NULL,
  b.tune = NULL,
  resK = FALSE,
  eta.star = NULL,
  len.imp = NULL,
  holdout = NULL
)
}
\arguments{
\item{niter}{number of iterations}

\item{nburn}{burn-in}

\item{y}{data}

\item{ycomplete}{complete data for validation}

\item{priors}{priors}

\item{K.true}{set number of hidden states}

\item{z.true}{set hidden state trajectories}

\item{lod}{limit of detection}

\item{missing}{indicator if there is missing data}

\item{gibbs_update}{indicator for gibbs sampling of Sigma or MH}

\item{tau2}{tuning param}

\item{a.tune}{tuning param}

\item{b.tune}{tuning param}

\item{resK}{tuning param}

\item{eta.star}{tuning param}

\item{len.imp}{number of imputations to save}

\item{holdout}{holdout data set for validation}
}
\value{
list
}
\description{
MVN model based clustering with fixed K and fixed state assignments
}
