% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismissing.R
\name{ismissing}
\alias{ismissing}
\title{function to indicate missing values}
\usage{
ismissing(x)
}
\arguments{
\item{x}{data}
}
\value{
matrix same size as x, NA's are 1, -Inf's are 2, observed data are 0
}
\description{
function to indicate missing values
}
