% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mciHMM.R
\name{mciHMM}
\alias{mciHMM}
\title{Fit covariate-dependent PSBP-iHMM to multiple time series}
\usage{
mciHMM(
  niter,
  nburn,
  y,
  X,
  rmlist = NULL,
  missing = FALSE,
  lod = NULL,
  len.imp = NULL,
  K.start = 12,
  priors = NULL,
  tau2 = NULL,
  a.tune = NULL,
  b.tune = NULL,
  resK = FALSE,
  eta.star = NULL,
  z.true = NULL,
  mu.true = NULL,
  ycomplete = NULL,
  holdout = NULL
)
}
\arguments{
\item{niter}{number of total iterations}

\item{nburn}{number of burn-in iterations}

\item{y}{list of time series data for each time series}

\item{X}{list, matrix of covariates for each time series}

\item{rmlist}{integer vector identifying repeated time series for the same subject with the same number (e.g. c(1,1,2,2,2,3,3))}

\item{missing}{logical; if TRUE then the data set y contains missing data, default is FALSE}

\item{lod}{list of lower limits of detection for p exposures for each time series}

\item{len.imp}{number of imputations to save. Imputations will be taken at equally spaced iterations between nburn and niter.}

\item{K.start}{starting number of hidden states, default is 12}

\item{priors}{list of priors}

\item{tau2}{variance tuning parameter for normal proposal in MH update of lower triangular elements in decomposition of Sigma}

\item{a.tune}{shape tuning parameter for inverse gamma proposal in MH update of diagonal elements in decomposition of Sigma}

\item{b.tune}{rate tuning parameter for inverse gamma proposal in MH update of diagonal elements in decomposition of Sigma}

\item{resK}{logical; if TRUE a resolvent kernel is used in MH update for lower triangular elements in decomposition of Sigma}

\item{eta.star}{resolvent kernel parameter, must be a real value greater than 1. In the resolvent kernel we take a random draw from the geometric distribution with mean (1-p)/p, eta.star = 1/p.}

\item{z.true}{list of true hidden states, if known}

\item{mu.true}{matrix of true exposure means for each true state, if known}

\item{ycomplete}{complete data, if available, for evaluating imputations}

\item{holdout}{list of indicators of missing type in holdout data set, 0 = observed, 1 = MAR, 2 = below LOD, for imputation validation purposes}
}
\value{
an object of type "iHMM"

a list with components
\itemize{
       \item z.save: list of estimated hidden states for each time series at each iteration
       \item K.save: list of estimated number of hidden states for each time series at each iteration
       \item ymar: matrix of imputed values for MAR data, number of rows equal to len.imp
       \item ylod: matrix of imputed values for data below LOD, number of rows equal to len.imp
       \item beta.save: list of posterior estimates of beta_k, state-specific regression coefficients in PSBP 
       \item beta_rm.save: list of posterior estimates of beta_sk, state-specific subject-specific regression coefficients in PSBP
       \item mu.save: list of posterior estimates of mu_k, state-specific means 
       \item hamming: posterior hamming distance between true and estimated states, if z.true is given
       \item mu.mse: mean squared error for estimated state-specific means, if mu.true is given
       \item mu.sse: sum of squared errors for estimated state-specific means, if mu.true is given
       \item mar.mse: mean squared error of MAR imputations, if ycomplete is given 
       \item lod.mse: mean squared error of imputations below LOD, if ycomplete is given 
       \item mar.sse: sum of squared errors of MAR imputations, if ycomplete is given 
       \item lod.sse: sum of squared errors of imputations below LOD, if ycomplete is given 
       \item mismat: list, each element is a matrix indicating types of missing data for each time series, 0 = complete, 1 = MAR, 2 = below LOD
       \item ycomplete: complete data
       \item MH.arate: average MH acceptance rate for lower triangular elements
       \item MH.lamrate: average MH acceptance rate for diagonal elements 
}
}
\description{
Fit covariate-dependent PSBP-iHMM to multiple time series
}
