% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveBeta.R
\name{modelAveBeta}
\alias{modelAveBeta}
\title{calculate model averaged estimates of harmonic trend (X %*% beta)}
\usage{
modelAveBeta(fit, zbest1, X)
}
\arguments{
\item{fit}{object of type "ihmm"}

\item{zbest1}{results from bestClusteriHMM}

\item{X}{covariates}
}
\value{
model averaged estimates of X %*% beta, state-specific covariate effect
}
\description{
calculate model averaged estimates of harmonic trend (X %*% beta)
}
