% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveMu.R
\name{modelAveMu}
\alias{modelAveMu}
\title{calculate model averaged estimates of mu}
\usage{
modelAveMu(fit, zbest1)
}
\arguments{
\item{fit}{object of type "ihmm"}

\item{zbest1}{results from best}
}
\value{
model averaged estimates of mu, state-specific means
}
\description{
calculate model averaged estimates of mu
}
