% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_Z.R
\name{recode_Z}
\alias{recode_Z}
\title{Relabel z to 1:length(z) while retaining order of z}
\usage{
recode_Z(z)
}
\arguments{
\item{z}{list of states}
}
\value{
recoded z
}
\description{
Relabel z to 1:length(z) while retaining order of z
}
